Successfully got mongodb to work.
Pulls data, parses it, saves the filtered data to a model then appends
it to a file. The model is then saved to MongoDB.

Problem: It appends the information each time the server is restarted.
Isn’t organized, just a big stack of a combination of old and new
pulls.